esperanto-tools-libs/requirements.txt
exhale==0.2.3
breathe==4.13.1
Jinja2==2.10.3
jsonschema==3.1.1
protobuf==3.6.1
PyYAML==5.1.2
Sphinx==2.2.1
sphinx-rtd-theme==0.4.3
